remove deprecated api & features
