--- packages/rath-client/src/components/backupModal/forms/index.tsx
@@ -1,8 +1,0 @@
-export interface IBackupFormHandler {
-    submit: (download: boolean) => Promise<boolean>;
-}
-
-export interface IBackupFormProps {
-    setBusy: (busy: boolean) => void;
-    setCanBackup: (able: boolean) => void;
-}
