+++ packages/rath-client/src/global.ts
@@ -3,7 +3,6 @@ export type Aggregator = 'sum' | 'mean' | 'count';
 export type OperatorType = 'sum' | 'mean' | 'count';
 
 export enum IDataSourceType {
-    CLOUD = 'kanaries_cloud',
     FILE = 'file',
     RESTFUL = 'restful',
     DATABASE = 'database',
