+++ packages/rath-client/src/pages/dashboard/dashboard-list.tsx
@@ -1,13 +1,12 @@
 import { DetailsList, DetailsRow, IColumn, IconButton, IDetailsRowProps, Layer, SelectionMode } from '@fluentui/react';
 import intl from 'react-intl-universal';
 import { observer } from 'mobx-react-lite';
-import { FC, useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import { FC, useCallback, useMemo, useRef, useState } from 'react';
 import styled from 'styled-components';
 import { useGlobalStore } from '../../store';
 import type { DashboardDocument } from '../../store/dashboardStore';
 import DocumentPreview from './document-preview';
 import { EditableCell } from './dashboard-homepage';
-import BackupDialog from './backup-dialog';
 
 
 const TableContainer = styled.div`
@@ -99,8 +98,7 @@ export interface DashboardListProps {
 }
 
 const DashboardList: FC<DashboardListProps> = ({ openDocument, pages }) => {
-    const { dashboardStore, userStore } = useGlobalStore();
-    const { loggedIn, cloudDataSourceMeta } = userStore;
+    const { dashboardStore } = useGlobalStore();
 
     const [sortMode, setSortMode] = useState<{
         key: Exclude<keyof FlatDocumentInfo, 'description' | 'index'>;
@@ -124,15 +122,6 @@ const DashboardList: FC<DashboardListProps> = ({ openDocument, pages }) => {
         lastModifyTime: p.info.lastModifyTime,
     })), [pages]);
 
-    // TODO: release dashboard upload feature
-    const canBackup = Boolean(loggedIn && cloudDataSourceMeta) && false;
-
-    const [selected, setSelected] = useState<FlatDocumentInfo | null>(null);
-
-    useEffect(() => {
-        setSelected(null);
-    }, [items]);
-
     const openDocumentRef = useRef(openDocument);
     openDocumentRef.current = openDocument;
 
@@ -141,7 +130,7 @@ const DashboardList: FC<DashboardListProps> = ({ openDocument, pages }) => {
             {
                 key: 'action',
                 name: '',
-                minWidth: canBackup ? 96 : 64,
+                minWidth: 64,
                 onRender(item) {
                     const { operators } = dashboardStore.fromPage(item['index']);
                     return (
@@ -149,7 +138,6 @@ const DashboardList: FC<DashboardListProps> = ({ openDocument, pages }) => {
                             <IconButton iconProps={{ iconName: 'BarChartVerticalEdit' }} onClick={() => openDocumentRef.current((item as FlatDocumentInfo).index)} />
                             <IconButton iconProps={{ iconName: 'Copy' }} onClick={operators.copy} />
                             {/* <IconButton iconProps={{ iconName: 'Download' }} onClick={operators.download} /> */}
-                            {canBackup && <IconButton iconProps={{ iconName: 'CloudUpload' }} onClick={() => setSelected(item)} />}
                             <IconButton iconProps={{ iconName: 'Delete', style: { color: '#f21044' } }} onClick={operators.remove} />
                         </ButtonGroup>
                     );
@@ -213,7 +201,7 @@ const DashboardList: FC<DashboardListProps> = ({ openDocument, pages }) => {
                 },
             },
         ];
-    }, [sortMode, dashboardStore, canBackup]);
+    }, [sortMode, dashboardStore]);
 
     const sortedItems = useMemo<typeof items>(() => {
         const flag = sortMode.direction === 'descending' ? -1 : 1;
@@ -291,13 +279,6 @@ const DashboardList: FC<DashboardListProps> = ({ openDocument, pages }) => {
                     </PreviewPopup>
                 )}
             </Layer>
-            {selected && (
-                <BackupDialog
-                    open={Boolean(selected)}
-                    onDismiss={() => setSelected(null)}
-                    value={selected}
-                />
-            )}
         </>
     );
 };
