+++ packages/rath-client/src/pages/dataSource/baseActions/dataOperations.tsx
@@ -130,14 +130,6 @@ const DataOperations: React.FC = () => {
                     dataSourceStore.setAllMutFieldsDisable(true);
                 },
             },
-            // {
-            //     key: 'backup',
-            //     text: intl.get('dataSource.operations.backup'),
-            //     iconProps: { iconName: 'CloudUpload' },
-            //     onClick: () => {
-            //         commonStore.setShowBackupModal(true);
-            //     },
-            // },
         ];
     }, [
         cleanMethod,
@@ -148,7 +140,6 @@ const DataOperations: React.FC = () => {
         exportDataAsJson,
         mutFields.length,
         exportDataAsRATHDS,
-        // commonStore,
     ]);
     return (
         <Cont>
