+++ packages/rath-client/src/pages/dataSource/selection/airtable/index.tsx
@@ -2,10 +2,9 @@ import { DefaultButton, PrimaryButton, Stack, TextField } from '@fluentui/react'
 import React, { useCallback, useState } from 'react';
 import intl from 'react-intl-universal'
 import { logDataImport } from '../../../../loggers/dataImport';
-import { DataSourceType, IMuteFieldBase, IRow } from '../../../../interfaces';
+import { IMuteFieldBase, IRow } from '../../../../interfaces';
 import { rawData2DataWithBaseMetas } from '../../utils';
 import { DataSourceTag, IDBMeta } from '../../../../utils/storage';
-import { useGlobalStore } from '../../../../store';
 import { fetchAllRecordsFromAirTable } from './utils';
 
 
@@ -16,7 +15,6 @@ interface AirTableSourceProps {
     onDataLoaded: (fields: IMuteFieldBase[], dataSource: IRow[], name: string, tag: DataSourceTag, withHistory?: IDBMeta | undefined) => void;
 }
 const AirTableSource: React.FC<AirTableSourceProps> = (props) => {
-    const { userStore } = useGlobalStore();
     const { onClose, onDataLoaded, onLoadingFailed, onStartLoading } = props;
     const [endPoint, setEndPoint] = useState<string>('');
     const [apiKey, setAPIKey] = useState<string>('');
@@ -44,15 +42,10 @@ const AirTableSource: React.FC<AirTableSourceProps> = (props) => {
                     dataSource: ds.dataSource.slice(0, 10),
                     size: ds.dataSource.length
                 });
-                userStore.saveDataSourceOnCloudOnlineMode({
-                    name,
-                    datasourceType: DataSourceType.AirTable,
-                    linkInfo,
-                });
                 onClose();
             })
             .catch(onLoadingFailed);
-    }, [onDataLoaded, onClose, onLoadingFailed, onStartLoading, endPoint, apiKey, tableID, tableName, viewName, userStore]);
+    }, [onDataLoaded, onClose, onLoadingFailed, onStartLoading, endPoint, apiKey, tableID, tableName, viewName]);
     return (
         <div>
             <Stack tokens={{ childrenGap: '4px' }} style={{ maxWidth: '300px' }}>
