+++ packages/rath-client/src/pages/dataSource/selection/index.tsx
@@ -29,13 +29,10 @@ interface SelectionProps {
 const Selection: React.FC<SelectionProps> = props => {
     const { show, onClose, onDataLoaded, loading, onStartLoading, onLoadingFailed, onDataLoading, setLoadingAnimation } = props;
 
-    const [dataSourceType, setDataSourceType] = useState<Exclude<IDataSourceType, IDataSourceType.CLOUD>>(IDataSourceType.LOCAL);
+    const [dataSourceType, setDataSourceType] = useState<IDataSourceType>(IDataSourceType.LOCAL);
     const dsTypeOptions = useDataSourceTypeOptions();
 
-    const formMap: Record<Exclude<IDataSourceType, IDataSourceType.CLOUD>, JSX.Element> = {
-        // [IDataSourceType.CLOUD]: (
-        //     <KanariesCloud setLoadingAnimation={setLoadingAnimation} />
-        // ),
+    const formMap: Record<IDataSourceType, JSX.Element> = {
         [IDataSourceType.FILE]: (
             <FileData onDataLoading={onDataLoading} onClose={onClose} onDataLoaded={onDataLoaded} onLoadingFailed={onLoadingFailed} toggleLoadingAnimation={setLoadingAnimation} />
         ),
@@ -72,7 +69,7 @@ const Selection: React.FC<SelectionProps> = props => {
                     selectedKey={dataSourceType}
                     onChange={(ev, option) => {
                         if (option) {
-                            setDataSourceType(option.key as Exclude<IDataSourceType, IDataSourceType.CLOUD>);
+                            setDataSourceType(option.key as IDataSourceType);
                         }
                     }}
                 />
