+++ packages/rath-client/src/store/commonStore.ts
@@ -7,9 +7,6 @@ import { VegaGlobalConfig } from '../queries/themes/config';
 import { destroyRathWorker, initRathWorker, rathEngineService } from '../services/index';
 import { transVegaSubset2Schema } from '../utils/transform';
 import { deepcopy } from '../utils';
-import { getMainServiceAddress } from '../utils/user';
-import { request } from '../utils/request';
-import { notify } from '../components/error';
 
 export type ErrorType = 'error' | 'info' | 'success';
 const TASK_TEST_MODE_COOKIE_KEY = 'task_test_mode';
@@ -34,20 +31,12 @@ export type ThemeListResponse = {
     pageIndex: number;
 };
 
-interface IWorkspaceSimpleInfo {
-    name: string;
-    organization: {
-        name: string;
-    };
-}
-
 export class CommonStore {
     public appKey: string = PIVOT_KEYS.connection;
     public computationEngine: string = COMPUTATION_ENGINE.webworker;
     public exploreMode: string = EXPLORE_MODE.comprehensive;
     public taskMode: ITaskTestMode = ITaskTestMode.local;
     public showStorageModal: boolean = false;
-    public showBackupModal: boolean = false;
     public showAnalysisConfig: boolean = false;
     public navMode: 'text' | 'icon' = 'text';
     public graphicWalkerSpec: Specification;
@@ -116,9 +105,6 @@ export class CommonStore {
     public setShowStorageModal(show: boolean) {
         this.showStorageModal = show;
     }
-    public setShowBackupModal(show: boolean) {
-        this.showBackupModal = show;
-    }
     public async setComputationEngine(engine: string) {
         try {
             destroyRathWorker();
@@ -136,53 +122,5 @@ export class CommonStore {
     }
     public async setExploreMode(mode: string) {
         this.exploreMode = mode;
-    }
-    public async getCloudThemes(userName: string): Promise<IThemeInfo[]> {
-        const workspaceUrl = getMainServiceAddress("/api/ce/simpleInfo/workspace");
-        let workspaceName: string;
-        try {
-            const result = await request.get<{ userName: string }, IWorkspaceSimpleInfo>(workspaceUrl, { userName });
-            workspaceName = result.name;
-        } catch (error) {
-            notify({
-                title: '[/api/ce/simpleInfo/workspace]',
-                type: 'error',
-                content: `${error}`,
-            });
-            return [];
-        }
-        const list: IThemeInfo[] = [];
-        const personalUrl = getMainServiceAddress('/api/ce/theme/list');
-        try {
-            const res = await request.get<{ workspaceName: string }, ThemeListResponse>(
-                personalUrl,
-                { workspaceName }
-            );
-            list.push(...res.list);
-        } catch (error) {
-            notify({
-                title: '[/api/ce/theme/list]',
-                type: 'error',
-                content: `${error}`,
-            });
-        }
-        const collectionUrl = getMainServiceAddress('/api/ce/theme/favorites/list');
-        try {
-            const res = await request.get<{}, ThemeListResponse>(collectionUrl);
-            list.push(...res.list);
-        } catch (error) {
-            notify({
-                title: '[/api/ce/theme/favorites/list]',
-                type: 'error',
-                content: `${error}`,
-            });
-        }
-        runInAction(() => {
-            for (const theme of list) {
-                // TODO: async config fetch
-                this.themes[theme.name] = theme.config;
-            }
-        });
-        return list;
     }
 }
