+++ packages/rath-client/src/store/userStore.ts
@@ -1,30 +1,32 @@
 import { makeAutoObservable, observable, runInAction } from 'mobx';
-import { TextWriter, ZipReader } from "@zip.js/zip.js";
-import {
-    DataSourceType,
-    type ICreateDashboardConfig,
-    type ICreateDatasetPayload,
-    type ICreateDatasetResult,
-    type ICreateDataSourcePayload,
-    type ICreateDataSourceResult,
-    type IDatasetData,
-    type IDatasetMeta,
-    type IDataSourceMeta,
-    type IDashboardDocumentInfo,
-    type ILoginForm,
-    type INotebook,
-    type IOrganization,
-    type IUserInfo,
-    type IWorkspace,
-} from '../interfaces';
 import { getMainServiceAddress } from '../utils/user';
 import { notify } from '../components/error';
 import { request } from '../utils/request';
-import { KanariesDatasetFilenameCloud } from '../constants';
-import { IKRFComponents, IParseMapItem } from '../utils/download';
-import { getGlobalStore } from '.';
-export * from '../interfaces';
 
+export interface ILoginForm {
+    userName: string;
+    password: string;
+    email: string;
+}
+
+export interface INotebook {
+    readonly id: string;
+    readonly name: string;
+    readonly size: number;
+    readonly createAt: number;
+    readonly downLoadURL: string;
+}
+
+export interface IWorkspace {
+    readonly id: string;
+    readonly name: string;
+}
+
+export interface IOrganization {
+    readonly name: string;
+    readonly id: string;
+    workspaces?: readonly IWorkspace[] | null | undefined;
+}
 
 interface ISignUpForm {
     userName: string;
@@ -36,14 +38,22 @@ interface ISignUpForm {
     invCode: string;
 }
 
+
+interface IUserInfo {
+    userName: string;
+    email: string;
+    eduEmail: string;
+    phone: string;
+    avatarURL: string;
+    organizations?: readonly IOrganization[] | undefined;
+}
+
 export default class UserStore {
     public login!: ILoginForm;
     public signup!: ISignUpForm;
     public initialized = false;
     public isLoggedIn = false;
     public info: IUserInfo | null = null;
-    public planLoading = false;
-    public saving = false;
     public get loggedIn() {
         return this.info !== null;
     }
@@ -51,10 +61,6 @@ export default class UserStore {
         return this.info?.userName ?? null;
     }
 
-    /** Storage meta on cloud. */
-    public cloudDataSourceMeta: IDataSourceMeta | null = null;
-    /** Dataset storage meta on cloud. */
-    public cloudDatasetMeta: IDatasetMeta | null = null;
     /** Organization name */
     public currentOrgName: string | null = null;
     /** Workspace name */
@@ -177,624 +183,8 @@ export default class UserStore {
         }
     }
 
-    public async getNotebooksAndDatasets(organizationName: string, workspaceName: string, forceUpdate = false) {
-        const which = this.info?.organizations?.find(org => org.name === organizationName)?.workspaces?.find(wsp => wsp.name === workspaceName);
-        if (!which) {
-            return null;
-        }
-        if (!forceUpdate && which.notebooks !== undefined) {
-            return null;
-        }
-        const listNotebookApiUrl = getMainServiceAddress('/api/ce/notebook/list');
-        const listDatasetApiUrl = getMainServiceAddress('/api/ce/dataset/list');
-        try {
-            const result = await Promise.allSettled([
-                request.get<{ workspaceName: string }, { notebookList: INotebook[] }>(listNotebookApiUrl, {
-                    workspaceName,
-                }),
-                request.get<{ workspaceName: string }, { datasetList: IDatasetMeta[] }>(listDatasetApiUrl, {
-                    workspaceName,
-                })
-            ]).then(([notebookRes, datasetRes]) => ({
-                notebooks: notebookRes.status === 'fulfilled' ? notebookRes.value.notebookList : [],
-                datasets: datasetRes.status === 'fulfilled' ? datasetRes.value.datasetList : [],
-                errInfo: [
-                    notebookRes.status === 'rejected' ? notebookRes.reason : null,
-                    datasetRes.status === 'rejected' ? datasetRes.reason : null,
-                ].filter(reason => reason !== null),
-            }));
-            for (const errInfo of result.errInfo) {
-                notify({
-                    title: '[getNotebooksAndDatasets]',
-                    type: 'error',
-                    content: `${errInfo}`,
-                });
-            }
-            if (result.errInfo.length > 0) {
-                return null;
-            }
-            runInAction(() => {
-                which.notebooks = result.notebooks;
-                which.datasets = result.datasets;
-            });
-            return result;
-        } catch (error) {
-            notify({
-                title: '[getNotebooksAndDatasets]',
-                type: 'error',
-                content: `${error}`,
-            });
-            return null;
-        }
-    }
-
-    public async uploadNotebook(workspaceName: string, file: File) {
-        const url = getMainServiceAddress('/api/ce/notebook');
-        try {
-            const { uploadUrl, id } = (await (await fetch(url, {
-                method: 'PUT',
-                credentials: 'include',
-                headers: {
-                    'Content-Type': 'application/json',
-                },
-                body: JSON.stringify({
-                    name: file.name,
-                    type: 0,    // TODO: customize type of upload workspace
-                    workspaceName,
-                    fileType: file.type,
-                    introduction: '',
-                    size: file.size,
-                }),
-            })).json()).data;
-            await fetch(uploadUrl, {
-                method: 'PUT',
-                body: file,
-            });
-            await request.post<{ id: string }, {}>(getMainServiceAddress('/api/ce/notebook/callback'), { id });
-            return true;
-        } catch (error) {
-            notify({
-                title: '[/api/ce/notebook]',
-                type: 'error',
-                content: `${error}`,
-            });
-            return false;
-        }
-    }
-
-    public async openDataset(dataset: IDatasetMeta): Promise<boolean> {
-        const { downloadUrl, datasource, id, workspace } = dataset;
-        try {
-            const data = await fetch(downloadUrl, { method: 'GET' });
-            if (!data.ok) {
-                throw new Error(data.statusText);
-            }
-            if (!data.body) {
-                throw new Error('Request got empty body');
-            }
-            const ok = await this.loadDataset(data.body, dataset.organization.name, workspace.name, datasource.id, id);
-            if (ok) {
-                const etUrl = getMainServiceAddress('/api/ce/tracing/normal');
-                try {
-                    await fetch(etUrl, {
-                        method: 'POST',
-                        credentials: 'include',
-                        headers: {
-                            'Content-Type': 'application/json',
-                        },
-                        body: JSON.stringify({
-                            eventTime: Date.now(),
-                            eventType: 'datasetUsage',
-                            eventLabel: 'rath',
-                            eventValue1: dataset.id,
-                        }),
-                    });
-                } catch (error) {
-                    console.warn('EventTrack error <report dataset use>', error);
-                }
-            }
-            return ok;
-        } catch (error) {
-            notify({
-                type: 'error',
-                title: '[openDataset]',
-                content: `${error}`,
-            });
-            return false;
-        }
-    }
-
-    protected async loadDataset(
-        body: ReadableStream<Uint8Array> | File,
-        organizationName: string | null,
-        workspaceName: string,
-        dataSourceId: string,
-        datasetId: string
-    ): Promise<boolean> {
-        const { dataSourceStore } = getGlobalStore();
-        try {
-            const zipReader = new ZipReader(body instanceof File ? body.stream() : body);
-            const file = (await zipReader.getEntries()).find(entry => entry.filename === KanariesDatasetFilenameCloud);
-            if (!file) {
-                throw new Error('Dataset file not found');
-            }
-            const writer = new TextWriter();
-            const dataset = JSON.parse(await file.getData(writer)) as IDatasetData;
-            const { meta } = dataset;
-            const dataSource = await this.fetchDataSource(workspaceName, dataSourceId, true);
-            if (dataSource && organizationName) {
-                this.setCloudDataSource(dataSource, organizationName, workspaceName);
-                const cloudMeta = await this.fetchDataset(workspaceName, datasetId);
-                if (cloudMeta) {
-                    this.setCloudDataset(cloudMeta);
-                    // comment is mutable only in the cloud data
-                    for (const item of cloudMeta.meta) {
-                        const f = meta.mutFields.find(f => f.fid === item.fid);
-                        if (f) {
-                            f.comment = item.comment;
-                        }
-                    }
-                    await dataSourceStore.loadBackupDataStore(dataset.data, dataset.meta);
-                } else {
-                    throw new Error('Cloud dataset not found');
-                }
-            }
-            return true;
-        } catch (error) {
-            notify({
-                type: 'error',
-                title: '[loadDataset]',
-                content: `${error}`,
-            });
-            return false;
-        }
-    }
-
-    public async fetchDataSource(workspaceName: string, dataSourceId: string, silent = false): Promise<IDataSourceMeta | null> {
-        const dataSourceApiUrl = getMainServiceAddress('/api/ce/datasource');
-        try {
-            const dataSourceDetail = await request.get<{
-                workspaceName: string;
-                datasourceId: string;
-            }, IDataSourceMeta>(dataSourceApiUrl, { datasourceId: dataSourceId, workspaceName });
-            return dataSourceDetail;
-        } catch (error) {
-            const info = {
-                type: 'error',
-                title: '[fetchDataSource]',
-                content: `${error}`,
-            } as const;
-            if (silent) {
-                console.warn(info);
-            } else {
-                notify(info);
-            }
-            return null;
-        }
-    }
-
-    protected async fetchDataset(workspaceName: string, datasetId: string): Promise<IDatasetMeta | null> {
-        const dataSourceApiUrl = getMainServiceAddress('/api/ce/dataset');
-        try {
-            const dataSourceDetail = await request.get<{
-                workspaceName: string;
-                datasetId: string;
-            }, Omit<IDatasetMeta, 'workspaceName'>>(dataSourceApiUrl, { datasetId, workspaceName });
-            return dataSourceDetail;
-        } catch (error) {
-            notify({
-                type: 'error',
-                title: '[fetchDataset]',
-                content: `${error}`,
-            });
-            return null;
-        }
-    }
-
-    public async openNotebook(downLoadURL: string) {
-        try {
-            const data = await fetch(downLoadURL, { method: 'GET' });
-            if (!data.ok) {
-                throw new Error(data.statusText);
-            }
-            if (!data.body) {
-                throw new Error('Request got empty body');
-            }
-            await this.loadNotebook(data.body);
-        } catch (error) {
-            notify({
-                title: '[download notebook]',
-                type: 'error',
-                content: `${error}`,
-            });
-        }
-    }
-
-    protected async loadNotebook(body: ReadableStream<Uint8Array> | File) {
-        try {
-            const zipReader = new ZipReader(body instanceof File ? body.stream() : body);
-            const result = await zipReader.getEntries();
-            const manifestFile = result.find((entry) => {
-                return entry.filename === "parse_map.json";
-            });
-            if (!manifestFile) {
-                throw new Error('Cannot find parse_map.json')
-            }
-            const writer = new TextWriter();
-            const manifest = JSON.parse(await manifestFile.getData(writer)) as {
-                items: IParseMapItem[];
-                version: string;
-            };
-            const { dataSourceStore, causalStore, dashboardStore, collectionStore } = getGlobalStore();
-            for await (const { name, key } of manifest.items) {
-                const entry = result.find(which => which.filename === name);
-                if (!entry || key === IKRFComponents.meta) {
-                    continue;
-                }
-                const w = new TextWriter();
-                try {
-                    const res = await entry.getData(w);
-                    switch (key) {
-                        case IKRFComponents.data: {
-                            const metaFile = manifest.items.find(item => item.type === IKRFComponents.meta);
-                            if (!metaFile) {
-                                break;
-                            }
-                            const meta = result.find(which => which.filename === metaFile.name);
-                            if (!meta) {
-                                break;
-                            }
-                            const wm = new TextWriter();
-                            const rm = await meta.getData(wm);
-                            await dataSourceStore.loadBackupDataStore(JSON.parse(res), JSON.parse(rm));
-                            break;
-                        }
-                        case IKRFComponents.collection: {
-                            collectionStore.loadBackup(JSON.parse(res));
-                            break;
-                        }
-                        case IKRFComponents.causal: {
-                            causalStore.load(JSON.parse(res));
-                            break;
-                        }
-                        case IKRFComponents.dashboard: {
-                            dashboardStore.loadAll(JSON.parse(res));
-                            break;
-                        }
-                        default: {
-                            break;
-                        }
-                    }
-                } catch (error) {
-                    notify({
-                        title: 'Load Notebook Error',
-                        type: 'error',
-                        content: `${error}`,
-                    });
-                    continue;
-                }
-            }
-        } catch (error) {
-            notify({
-                title: 'Load Notebook Error',
-                type: 'error',
-                content: `${error}`,
-            });
-        }
-    }
-
-    public async openDashboardTemplates(dashboards: IDashboardDocumentInfo[]): Promise<boolean> {
-        const { dashboardStore } = getGlobalStore();
-        for await (const dashboard of dashboards) {
-            try {
-                const dashboardData = await fetch(dashboard.downloadUrl, { method: 'GET' });
-                if (!dashboardData.ok) {
-                    throw new Error(dashboardData.statusText);
-                }
-                if (!dashboardData.body) {
-                    throw new Error('Request got empty body');
-                }
-                const res = await dashboardData.json() as Parameters<(typeof dashboardStore)['loadPage']>[0];
-                dashboardStore.loadPage(res, dashboard);
-            } catch (error) {
-                notify({
-                    type: 'error',
-                    title: '[fetchDashboardTemplate]',
-                    content: `${error}`,
-                });
-                return false;
-            } 
-        }
-        return true;
-    }
-
-    public async uploadDashboard(workspaceName: string, file: File, config: ICreateDashboardConfig): Promise<boolean> {
-        const { dataSourceStore } = getGlobalStore();
-        const { fieldMetas } = dataSourceStore;
-        if (!this.cloudDataSourceMeta) {
-            notify({
-                type: 'error',
-                title: '[uploadDashboard]',
-                content: 'Data source is not uploaded',
-            });
-            return false;
-        }
-        const { id: datasourceId } = this.cloudDataSourceMeta;
-        const createDashboardApiUrl = getMainServiceAddress('/api/ce/dashboard');
-        const reportUploadSuccessApiUrl = getMainServiceAddress('/api/ce/upload/callback');
-        try {
-            const respond = await fetch(createDashboardApiUrl, {
-                method: 'POST',
-                headers: {
-                    'Content-Type': 'application/json; charset=utf-8',
-                },
-                body: JSON.stringify({
-                    datasourceId,
-                    datasetId: config.dashboard.bindDataset && this.cloudDatasetMeta ? this.cloudDatasetMeta.id : undefined,
-                    workspaceName,
-                    name: config.dashboard.name,
-                    description: config.dashboard.description,
-                    dashboardTemplate: {
-                        name: config.dashboardTemplate.name,
-                        description: config.dashboardTemplate.description,
-                        meta: fieldMetas.map(f => ({
-                            fId: f.fid,
-                            description: config.dashboardTemplate.fieldDescription[f.fid] || f.name || '',
-                            analyticType: f.analyticType,
-                            semanticType: f.semanticType,
-                        })),
-                        size: file.size,
-                        type: config.dashboardTemplate.publish ? 1 : 2,
-                        cover: config.dashboardTemplate.cover ? {
-                            name: config.dashboardTemplate.cover.name,
-                            size: config.dashboardTemplate.cover.size,
-                            type: config.dashboardTemplate.cover.type,
-                        } : undefined,
-                    },
-                }),
-            });
-            if (!respond.ok) {
-                throw new Error(respond.statusText);
-            }
-            const res = await respond.json() as {
-                cover?: { uploadUrl: string; storage_id: string };
-                dashboardTemplate: { uploadUrl: string; storageId: string };
-            };
-            if (config.dashboardTemplate.cover && res.cover?.uploadUrl) {
-                const coverUploadRes = await fetch(res.cover.uploadUrl, {
-                    method: 'PUT',
-                    body: file,
-                });
-                if (!coverUploadRes.ok) {
-                    throw new Error(`Failed to upload cover: ${coverUploadRes.statusText}`);
-                }
-                await request.get<{ storageId: string; status: 1 }, { downloadUrl: string }>(
-                    reportUploadSuccessApiUrl, { storageId: res.cover.storage_id, status: 1 }
-                );
-            }
-            const templateUploadRes = await fetch(res.dashboardTemplate.uploadUrl, {
-                method: 'PUT',
-                body: file,
-            });
-            if (!templateUploadRes.ok) {
-                throw new Error(`Failed to upload file: ${templateUploadRes.statusText}`);
-            }
-            await request.get<{ storageId: string; status: 1 }, { downloadUrl: string }>(
-                reportUploadSuccessApiUrl, { storageId: res.dashboardTemplate.storageId, status: 1 }
-            );
-            return true;
-        } catch (error) {
-            notify({
-                type: 'error',
-                title: '[uploadDashboard]',
-                content: `${error}`,
-            });
-            return false;
-        }
-    }
-
-    public async saveDataSourceOnCloudOfflineMode(
-        params: ICreateDataSourcePayload<'offline'>,
-        file: File,
-    ): Promise<{ id: string; downloadUrl: string } | null> {
-        const send = async () => {
-            if (!this.workspace || !this.organization || this.uploadingDataSource) {
-                return null;
-            }
-            runInAction(() => {
-                this.uploadingDataSource = true;
-            });
-            const createDataSourceApiUrl = getMainServiceAddress('/api/ce/datasource');
-            const reportUploadSuccessApiUrl = getMainServiceAddress('/api/ce/upload/callback');
-            try {
-                const createDataSourceApiRes = await request.post<typeof params & { workspaceName: string }, ICreateDataSourceResult<'offline'>>(
-                    createDataSourceApiUrl, {
-                        ...params,
-                        workspaceName: this.workspace.name,
-                    }
-                );
-                if (params.datasourceType === DataSourceType.File) {
-                    const fileUploadRes = await fetch(createDataSourceApiRes.fileInfo.uploadUrl, {
-                        method: 'PUT',
-                        body: file,
-                    });
-                    if (!fileUploadRes.ok) {
-                        throw new Error(`Failed to upload file: ${fileUploadRes.statusText}`);
-                    }
-                    await request.get<{ storageId: string; status: 1 }, { downloadUrl: string }>(
-                        reportUploadSuccessApiUrl, { storageId: createDataSourceApiRes.fileInfo.storageId, status: 1 }
-                    );
-                }
-                const dataSource = await this.fetchDataSource(this.workspace.name, createDataSourceApiRes.id);
-                if (!dataSource) {
-                    throw new Error('Data source not existed');
-                }
-                this.setCloudDataSource(dataSource, this.organization.name, this.workspace.name);
-                const fileUploadRes = await fetch(createDataSourceApiRes.fileInfo.uploadUrl, {
-                    method: 'PUT',
-                    body: file,
-                });
-                if (!fileUploadRes.ok) {
-                    throw new Error(`Failed to upload file: ${fileUploadRes.statusText}`);
-                }
-                const reportUploadSuccessApiRes = await request.get<{ storageId: string; status: 1 }, { downloadUrl: string }>(
-                    reportUploadSuccessApiUrl, { storageId: createDataSourceApiRes.fileInfo.storageId, status: 1 }
-                );
-                runInAction(() => {
-                    this.uploadingDataSource = false;
-                });
-                return {
-                    id: createDataSourceApiRes.id,
-                    downloadUrl: reportUploadSuccessApiRes.downloadUrl,
-                };
-            } catch (error) {
-                notify({
-                    type: 'error',
-                    title: '[saveDataSourceOnCloud]',
-                    content: `${error}`,
-                });
-                runInAction(() => {
-                    this.uploadingDataSource = false;
-                });
-                return null;
-            }
-        };
-        if (this.organization && this.workspace && !this.uploadingDataSource) {
-            return send();
-        }
-        this.uploadDataSource = () => {
-            if (this.organization && this.workspace && !this.uploadingDataSource) {
-                send().then(res => {
-                    if (res) {
-                        this.uploadDataSource = null;
-                    }
-                });
-            }
-        };
-        return null;
-    }
-
-    public async saveDataSourceOnCloudOnlineMode(
-        params: ICreateDataSourcePayload<'online'>,
-    ): Promise<{ id: string } | null> {
-        const send = async () => {
-            if (!this.workspace || !this.organization || this.uploadingDataSource) {
-                return null;
-            }
-            runInAction(() => {
-                this.uploadingDataSource = true;
-            });
-            const createDataSourceApiUrl = getMainServiceAddress('/api/ce/datasource');
-            try {
-                const createDataSourceApiRes = await request.post<typeof params & { workspaceName: string }, ICreateDataSourceResult<'online'>>(
-                    createDataSourceApiUrl, {
-                        ...params,
-                        workspaceName: this.workspace.name,
-                    }
-                );
-                const dataSource = await this.fetchDataSource(this.workspace.name, createDataSourceApiRes.id);
-                if (!dataSource) {
-                    throw new Error('Data source not existed');
-                }
-                this.setCloudDataSource(dataSource, this.organization.name, this.workspace.name);
-                runInAction(() => {
-                    this.uploadingDataSource = false;
-                });
-                return { id: createDataSourceApiRes.id };
-            } catch (error) {
-                notify({
-                    type: 'error',
-                    title: '[saveDataSourceOnCloud]',
-                    content: `${error}`,
-                });
-                runInAction(() => {
-                    this.uploadingDataSource = false;
-                });
-                return null;
-            }
-        };
-        if (this.organization && this.workspace && !this.uploadingDataSource) {
-            return send();
-        }
-        this.uploadDataSource = () => {
-            if (this.organization && this.workspace && !this.uploadingDataSource) {
-                send().then(res => {
-                    if (res) {
-                        this.uploadDataSource = null;
-                    }
-                });
-            }
-        };
-        return null;
-    }
-
-    public async saveDatasetOnCloud(payload: ICreateDatasetPayload, file: File) {
-        const { userStore } = getGlobalStore();
-        const createDatasetApiUrl = getMainServiceAddress('/api/ce/dataset');
-        const reportUploadSuccessApiUrl = getMainServiceAddress('/api/ce/upload/callback');
-        try {
-            const respond = await fetch(createDatasetApiUrl, {
-                method: 'POST',
-                headers: {
-                    'Content-Type': 'application/json; charset=utf-8',
-                },
-                credentials: 'include',
-                body: JSON.stringify(payload),
-            });
-            if (!respond.ok) {
-                throw new Error(respond.statusText);
-            }
-            const createDatasetApiRes = await respond.json() as (
-                | { success: false; message: string }
-                | { success: true; data: ICreateDatasetResult }
-            );
-            if (!createDatasetApiRes.success) {
-                throw new Error(createDatasetApiRes.message);
-            }
-            const fileUploadRes = await fetch(createDatasetApiRes.data.uploadUrl, {
-                method: 'PUT',
-                body: file,
-            });
-            if (!fileUploadRes.ok) {
-                throw new Error(`Failed to upload file: ${fileUploadRes.statusText}`);
-            }
-            const reportUploadSuccessApiRes = await request.get<{ storageId: string; status: 1 }, { downloadUrl: string }>(
-                reportUploadSuccessApiUrl, { storageId: createDatasetApiRes.data.storageId, status: 1 }
-            );
-            const dataset = await userStore.fetchDataset(payload.workspaceName, createDatasetApiRes.data.datasetId);
-            if (!dataset) {
-                throw new Error('Dataset not existed');
-            }
-            this.setCloudDataset(dataset);
-            return reportUploadSuccessApiRes;
-        } catch (error) {
-            notify({
-                type: 'error',
-                title: '[saveDatasetOnCloud]',
-                content: `${error}`,
-            });
-            return null;
-        }
-    }
-
-    public setCloudDataset(dataset: IDatasetMeta) {
-        this.cloudDatasetMeta = dataset;
-        this.currentOrgName = dataset.organization.name;
-        this.currentWspName = dataset.workspace.name;
-    }
-
-    public setCloudDataSource(dataSource: IDataSourceMeta, organizationName?: string, workspaceName?: string) {
-        this.cloudDataSourceMeta = dataSource;
-        if (organizationName) {
-            this.currentOrgName = organizationName;
-        }
-        if (workspaceName) {
-            this.currentWspName = workspaceName;
-        }
-    }
-
-    public setOrgId(orgId: string | null): boolean {
-        const org = this.info?.organizations?.find(which => which.id === orgId);
+    public setOrgName(orgName: string | null): boolean {
+        const org = this.info?.organizations?.find(which => which.name === orgName);
         if (!org) {
             this.currentOrgName = null;
             return false;
